roughly splits Builder into three classes and other minors for #25
